+++ docs/source/change_log.rst
@@ -16,6 +16,8 @@ Version 1.8.0
 * Support the client on OS X by using Docker
 * Support for issuing certificates  with acme while the server is running
 * Add a wrapping tool for certbot to make the process easier
+* Updated `tools/cx_freeze.py` to build KingPhisher client in Python 3.4
+* Updated documentation for windows build.
 
 Version 1.7.1
 ^^^^^^^^^^^^^